requirements.txt
streamlit
pandas
requests
fastapi
uvicorn
tensorflow
google-cloud-storage
google-auth
google-auth-oauthlib
tqdm
numpy

